% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar..capas.sig.R
\docType{data}
\name{departamento}
\alias{departamento}
\title{departamento}
\format{
Un objeto Simple Feature de 529 observaciones y 11 variables
}
\usage{
data(departamento)
}
\description{
División político administrativa de segundo orden que incluye partido y comuna.
}
\details{
Atributos:
\itemize{
  \item{gid}: id
  \item{fna}: Nombre geográfico
  \item{gna}: Término genérico
  \item{nam}: Término específico
  \item{in1}: Código INDEC
  \item{fdc}: Fuente de captura
  \item{sag}: Autoridad de fuente
  \item{provincia_gid}: id (provincia)
  \item{provincia_fna}: Nombre geográfico (provincia)
  \item{provincia_nam}: Término específico (provincia)
}
}
\references{
\url{https://www.ign.gob.ar/NuestrasActividades/InformacionGeoespacial/CapasSIG}
}
\author{
Instituto Geográfico Nacional
}
\keyword{datasets}
