% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar..capas.sig.R
\docType{data}
\name{doscientas_millas_sector_antartico}
\alias{doscientas_millas_sector_antartico}
\title{doscientas_millas_sector_antartico}
\format{
Un objeto Simple Feature de 1 observación y 4 variables
}
\usage{
data(doscientas_millas_sector_antartico)
}
\description{
Zona marítima que se extiende más allá del límite exterior del mar territorial,
hasta una distancia de 200 millas marinas a partir de la línea de base.
Sector antártico.
}
\details{
Atributos:
\itemize{
  \item{SAG}: Autoridad de fuente
  \item{FDC}: Fuente de captura
}
}
\references{
\url{https://www.ign.gob.ar/NuestrasActividades/InformacionGeoespacial/CapasSIG}
}
\author{
Instituto Geográfico Nacional
}
\keyword{datasets}
