% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar..capas.sig.R
\docType{data}
\name{municipio}
\alias{municipio}
\title{Municipio}
\format{
Un objeto Simple Feature de 1808 observaciones y 8 variables
}
\usage{
data(municipio)
}
\description{
División político administrativa de tercer orden que incluye comunas,
juntas vecinales y demás formas de gobierno local.
}
\details{
Atributos:
\itemize{
  \item{FNA}: Nombre geográfico
  \item{GNA}: Término genérico
  \item{NAM}: Término específico
  \item{IN1}: Código INDEC
  \item{FDC}: Fuente de captura
  \item{SAG}: Autoridad de fuente
}
}
\references{
\url{https://www.ign.gob.ar/NuestrasActividades/InformacionGeoespacial/CapasSIG}
}
\author{
Instituto Geográfico Nacional
}
\keyword{datasets}
