% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar..capas.sig.R
\docType{data}
\name{pais}
\alias{pais}
\title{pais}
\format{
Un objeto Simple Feature de 1 observacióne y 8 variables
}
\usage{
data(pais)
}
\description{
País se define como una nación jurídicamente organizada. En la entidad se
incluyen los límites de la República Argentina.
}
\details{
Atributos:
\itemize{
  \item{fna}: Nombre geográfico
  \item{GNA}: Término genérico
  \item{NAM}: Término específico
  \item{FDC}: Fuente de captura
  \item{SAG}: Autoridad de fuente
}
}
\references{
\url{https://www.ign.gob.ar/NuestrasActividades/InformacionGeoespacial/CapasSIG}
}
\author{
Instituto Geográfico Nacional
}
\keyword{datasets}
