% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar..capas.sig.R
\docType{data}
\name{provincia}
\alias{provincia}
\title{provincia}
\format{
Un objeto Simple Feature de 24 observaciones y 8 variables
}
\usage{
data(provincia)
}
\description{
País se define como una nación jurídicamente organizada. En la entidad se
incluyen los límites de la República Argentina.
}
\details{
Atributos:
\itemize{
  \item{fna}: Nombre geográfico
  \item{gna}: Término genérico
  \item{nam}: Término específico
  \item{in1}: Código INDEC
  \item{fdc}: Fuente de captura
  \item{sag}: Autoridad de fuente
}
}
\references{
\url{https://www.ign.gob.ar/NuestrasActividades/InformacionGeoespacial/CapasSIG}
}
\author{
Instituto Geográfico Nacional
}
\keyword{datasets}
